Subjectの詳細ページをクラス名に依存しないように修正
